adds id to book response

+++ src/main/java/bookworld_api/entities/Book.java
@@ -2,7 +2,7 @@ package bookworld_api.entities;
 
 public class Book {
 
-  private int id;
+  private long id;
   private String title;
   private String author;
   private String country;
@@ -20,7 +20,7 @@ public class Book {
     this.thumbnail = thumbnail;
   }
 
-  public Book(int id, String title, String author, String country, String description,
+  public Book(long id, String title, String author, String country, String description,
       String thumbnail) {
     this.id = id;
     this.title = title;
@@ -30,7 +30,7 @@ public class Book {
     this.thumbnail = thumbnail;
   }
 
-  public int getId() {
+  public long getId() {
     return id;
   }
