+++ src/main/java/bookworld_api/repositories/PostgresBookRepository.java
@@ -44,7 +44,7 @@ public class PostgresBookRepository implements BookRepository {
     ResultSet resultSet = preparedStatement.executeQuery();
     if (resultSet.next()) {
       return new Book(
-          resultSet.getInt("id"), resultSet.getString("title"),
+          resultSet.getLong("id"), resultSet.getString("title"),
           resultSet.getString("author"), resultSet.getString("country"),
           resultSet.getString("description"), resultSet.getString("thumbnail")
       );
@@ -58,7 +58,7 @@ public class PostgresBookRepository implements BookRepository {
     ResultSet resultSet = psqlConnection().createStatement().executeQuery(BOOKS_LIST);
     while (resultSet.next()) {
       books.add(new Book(
-          resultSet.getInt("id"), resultSet.getString("title"),
+          resultSet.getLong("id"), resultSet.getString("title"),
           resultSet.getString("author"), resultSet.getString("country"),
           resultSet.getString("description"), resultSet.getString("thumbnail")
       ));
